refactor: change `Route` type from `ArrayList<>` to `List<>`
the reasoning is (a) that this leaks the implementation detail into the
public API and (2) the the call site usage, e.g.
```
dispatch(SetRouteAction(arrayListOf("route-a", "route-b")
// vs.
dispatch(SetRouteAction(listOf("route-a", "route-b")
```

This does not break API contract, all code using the current version of
ReKotlin-router still compiles fine, because every instance of
`ArrayList` is also a `List`.